+++ src/docs/HISTORY.md
@@ -7,6 +7,8 @@ v1.7.4
 - UI: Fixed clipping of file names on the Compare Panel.
 - UI/Core: Improved thumbnails performance of all files, particularly .otio files.
 
+- Edit: .otioz files are now loaded in memory.
+- Edit: Added the ability to edit .otioz files.
 - Edit: Adding clips to a timeline with dissolves now keeps the dissolves intact instead of removing them.
 - Edit: Improved the performance of editing .otio files when dragging clips to the Timeline Viewport.  This removes the clunky file switching that used to happen.
 - Edit: The Playlist panel has been removed.  Now all clip concatenation operations happen directly by dragging clips to the Timeline Viewport.